third_party/WebKit/LayoutTests/platform/android/svg/animations/animate-text-nested-transforms-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

This is some text.
test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL rootSVGElement.getBBox().x should be close to 117. Was 115.01837921142578.
FAIL rootSVGElement.getBBox().y should be close to -160. Was -154.0113525390625.
PASS rootSVGElement.getBBox().x is 367
PASS rootSVGElement.getBBox().y is 550
PASS successfullyParsed is true

TEST COMPLETE
